Object Manager | Removing Dependancies
Features: (1) Created a new Object Manager class that is responsible for allocating and deallocating memory. Planing on plugging a memory pool to work with it as well. (2) Started removing dependancies between the Grid and objects, such as creating and deleting objects. Planing on simplifying how the grid works, making its' purpose more clear.