Started Seperating Grid from Lifespan
Features: (1) Using the visitor pattern I made some helper functions that allow us to pass information between the grid and other objects without needing to inherit from them. (2) Hooked up the deleting and creation of objects to only partially use the grid, simply calling a function that calls another function. All that is left is to remove the middle function from the grid and we should be okay.